perf: keep colorjs.io out of the renderers and drop the brightened color cache
Importing the hex helpers from colors/color.ts pulled colorjs.io into every
renderer and therefore into the export worker bundle (+29 kB). The plain hex
helpers now live in colors/hex.ts with no colorjs.io dependency.

The horizontal piano roll cached brightened colors by the raw float flash
intensity, which changes every frame, so the cache grew without bound during
playback. Brightening is a three-channel parse and is cheap enough to compute
each frame.
